fixes rust-analyzer install, do not reinstall if already present

+++ setup/lsp.sh
@@ -10,25 +10,35 @@ source deps.sh
 # exists rust-analyzer-preview || echo -e "${RED}please make sure $HOME/.cargo/bin is in path"
 
 # rust
-rustup=$(ensure_rustup)
-$rustup component add rust-src
-curl -L https://github.com/rust-analyzer/rust-analyzer/releases/latest/download/rust-analyzer-linux \
-	-o ~/.local/bin/rust-analyzer
-chmod +x $HOME/.local/bin/rust-analyzer
-exists rust-analyzer || echo -e "${RED}please make sure $HOME/.local/bin is in path"
+if ! exists rust-analyzer; then
+	rustup=$(ensure_rustup)
+	$rustup component add rust-src
+	release="rust-analyzer-x86_64-unknown-linux-gnu.gz"
+	base_url="https://github.com/rust-analyzer/rust-analyzer/releases/latest/download/"
+	local_path="$HOME/.local/bin/rust-analyzer"
+	curl -L "$base_url$release" | gunzip -c - > $local_path
+	chmod +x $local_path
+	exists rust-analyzer || echo -e "${RED}please make sure $HOME/.local/bin is in path"
+fi
 
 # python
-pip=$(ensure_pip)
-$pip install --user --upgrade 'python-lsp-server[all]'
+if ! exists pylsp; then
+	pip=$(ensure_pip)
+	$pip install --user --upgrade 'python-lsp-server[all]'
+fi
 
-# # latex
-cargo=$(ensure_cargo)
-$cargo install --git https://github.com/latex-lsp/texlab.git --locked
-exists texlab || echo -e "${RED}please make sure $HOME/.cargo/bin is in path"
+# latex
+if ! exists texlab; then
+	cargo=$(ensure_cargo)
+	$cargo install --git https://github.com/latex-lsp/texlab.git --locked
+	exists texlab || echo -e "${RED}please make sure $HOME/.cargo/bin is in path"
+fi
 
-# # bash
-npm=$(ensure_npm)
-make -p ~/.local/bin
-$npm config set prefix '~/.local/'
-$npm install -g bash-language-server
-exists bash-language-server || echo -e "${RED}please make sure $($npm bin) is in path"
+# bash
+if ! exists bash-language-server; then
+	npm=$(ensure_npm)
+	make -p ~/.local/bin
+	$npm config set prefix '~/.local/'
+	$npm install -g bash-language-server
+	exists bash-language-server || echo -e "${RED}please make sure $($npm bin) is in path"
+fi
